pyproject.toml
[project]
name = "callmind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=0.4.24",
    "fastapi>=0.116.1",
    "google-genai>=1.26.0",
    "google-generativeai>=0.8.5",
    "langchain-chroma>=0.1.2",
    "mcp>=1.12.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.28.0",
    "uvicorn>=0.35.0",
]
